TicketingTransactionService tests

+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/service/v2/TicketingTransactionService.java
@@ -30,9 +30,11 @@ import com.palantir.atlasdb.transaction.service.AbstractKeyValueServiceBackedTra
  *
  * We divide the first PARTITIONING_QUANTUM timestamps among the first ROW_PER_QUANTUM rows.
  * We aim to distribute start timestamps as evenly as possible among these rows as numbers increase, by taking the
- * least significant bits of the timestamp and using that as the row number. We store the row name as a little-endian
- * representation of the row number to ensure even distribution in key-value services that rely on consistent hashing
- * or similar mechanisms for partitioning.
+ * least significant bits of the timestamp and using that as the row number. For example, we might store timestamps
+ * 1, ROW_PER_QUANTUM + 1, 2 * ROW_PER_QUANTUM + 1 etc. in the same row.
+ *
+ * We store the row name as a little-endian representation of the row number to ensure even distribution in key-value
+ * services that rely on consistent hashing or similar mechanisms for partitioning.
  *
  * We also use a delta encoding for the commit timestamp as these differences are expected to be small.
  *
@@ -42,7 +44,7 @@ import com.palantir.atlasdb.transaction.service.AbstractKeyValueServiceBackedTra
 public class TicketingTransactionService extends AbstractKeyValueServiceBackedTransactionService {
     // DO NOT change the following without a transactions table migration!
     public static final long PARTITIONING_QUANTUM = 400_000_000;
-    public static final long ROWS_PER_QUANTUM = 256;
+    public static final int ROWS_PER_QUANTUM = 256;
 
     public TicketingTransactionService(KeyValueService keyValueService) {
         super(keyValueService);
